requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
psycopg2
python-jose
passlib
pydantic[email]
bcrypt
python-multipart
emails
jinja2
pylint
aioredis[hiredis]
flake8
autoflake
fastapi-cache2[redis]
slack_sdk
aiohttp
typer
pytest
pytest-asyncio
paramiko
pyyaml